Separate and organize code into different modules
