+++ MohammadBallal/app/main.py
@@ -1,0 +1,6 @@
+from fastapi import FastAPI
+from app.routes import router
+
+app = FastAPI()
+
+app.include_router(router)
